add updateRestaurantDto

+++ backend/src/restaurants/restaurants.service.ts
@@ -1,7 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Restaurant } from './schemas/restaurant.schema';
 import * as mongoose from 'mongoose';
+import { Query } from 'express-serve-static-core';
 
 @Injectable()
 export class RestaurantsService {
@@ -10,8 +15,22 @@ export class RestaurantsService {
     private restaurantModel: mongoose.Model<Restaurant>,
   ) {}
 
-  async findAll(): Promise<Restaurant[]> {
-    const restautants = await this.restaurantModel.find();
+  async findAll(query: Query): Promise<Restaurant[]> {
+    const page = Number(query.page) || 1;
+    const limit = 8;
+    const skip = (page - 1) * limit;
+    const keyword = query.keyword
+      ? {
+          name: {
+            $regex: query.keyword,
+            $options: 'i',
+          },
+        }
+      : {};
+    const restautants = await this.restaurantModel
+      .find(keyword)
+      .limit(limit)
+      .skip(skip);
     return restautants;
   }
   async createNew(restaurant: Restaurant): Promise<Restaurant> {
@@ -19,6 +38,10 @@ export class RestaurantsService {
     return res;
   }
   async findById(id: string): Promise<Restaurant> {
+    const isValidId = mongoose.isValidObjectId(id);
+    if (!isValidId) {
+      throw new BadRequestException('Please enter correct ID');
+    }
     const res = await this.restaurantModel.findById(id);
     if (!res) {
       throw new NotFoundException('Restaurant not found.');
@@ -26,12 +49,20 @@ export class RestaurantsService {
     return res;
   }
   async updateById(id: string, restaurant: Restaurant): Promise<Restaurant> {
+    const isValidId = mongoose.isValidObjectId(id);
+    if (!isValidId) {
+      throw new BadRequestException('Please enter correct ID');
+    }
     return await this.restaurantModel.findByIdAndUpdate(id, restaurant, {
       new: true,
       runValidators: true,
     });
   }
   async deleteById(id: string): Promise<Restaurant> {
+    const isValidId = mongoose.isValidObjectId(id);
+    if (!isValidId) {
+      throw new BadRequestException('Please enter correct ID');
+    }
     return await this.restaurantModel.findByIdAndDelete(id);
   }
 }
